requirements.txt
langchain
langchain-community
beautifulsoup4
chromadb
gradio
groq
ollama
python-dotenv